+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2024-10-18
+
+### Fixed
+
+- Fixed ProtectedRoute component to atomic structure
+- Added routes to atomic structure
+
 ## [0.0.2] - 2024-10-18
 
 ### Fixed